Time: 10 ms (8.73%), Space: 34 MB (43.41%) - LeetHub

+++ 1354-construct-target-array-with-multiple-sums/1354-construct-target-array-with-multiple-sums.cpp
@@ -20,7 +20,7 @@ public:
         while(!pq.empty() && pq.top() != 1)
         {
             auto top = pq.top(); pq.pop();
-            int R = sum - top;
+            long long R = sum - top;
             if(R == 1) return true;
             if(R < 1 || top <= R || top % R == 0) return false;
             int prev = top % R;
